+++ benchopt/cli/completion.py
@@ -3,6 +3,14 @@ from pathlib import Path
 
 from benchopt.benchmark import Benchmark
 from benchopt.utils.safe_import import skip_import
+from benchopt.utils.conda_env_cmd import list_conda_envs
+
+
+def propose_from_list(candidates, incomplete):
+    proposals = [c for c in candidates if str(c).startswith(incomplete)]
+    if len(proposals) > 0:
+        return proposals
+    return [c for c in candidates if incomplete in str(c)]
 
 
 def get_benchmark(ctx, args, incomplete):
@@ -17,19 +25,17 @@ def get_benchmark(ctx, args, incomplete):
 
     # List all sub directory
     all_dirs = [d for d in incomplete_path.glob('*') if d.is_dir()]
+    all_benchmarks = [
+        b for b in all_dirs if (Path(b) / "objective.py").exists()
+    ]
 
     # First try to list benchmarks that match the incomplete pattern.
-    benchmarks = [
-        b for b in all_dirs
-        if (Path(b) / "objective.py").exists() and incomplete in str(b)
-    ]
-    if len(benchmarks) > 0:
-        return benchmarks
+    proposed_benchmarks = propose_from_list(all_benchmarks, incomplete)
+    if len(proposed_benchmarks) > 0:
+        return proposed_benchmarks
 
     # Else do completion with sub-directories.
-    matching_dirs = [
-        b for b in all_dirs if incomplete in str(b)
-    ]
+    matching_dirs = propose_from_list(all_dirs, incomplete)
     if len(matching_dirs) == 1:
         # If only one matches, complete the folder name and continue completion
         # from here.
@@ -54,7 +60,7 @@ def get_solvers(ctx, args, incomplete):
         return [("", 'Benchmark has not been provided before')]
     solvers = benchmark.list_benchmark_solver_names()
     solvers = [s.lower() for s in solvers]
-    return [s for s in solvers if incomplete.lower() in s]
+    return propose_from_list(solvers, incomplete.lower())
 
 
 def get_datasets(ctx, args, incomplete):
@@ -65,7 +71,7 @@ def get_datasets(ctx, args, incomplete):
         return [("", 'Benchmark has not been provided before')]
     datasets = benchmark.list_benchmark_dataset_names()
     datasets = [d.lower() for d in datasets]
-    return [d for d in datasets if incomplete.lower() in d]
+    return propose_from_list(datasets, incomplete.lower())
 
 
 def get_output_files(ctx, args, incomplete):
@@ -75,6 +81,11 @@ def get_output_files(ctx, args, incomplete):
     if benchmark is None:
         return [("", 'Benchmark has not been provided before')]
     output_folder = benchmark.get_output_folder()
-    return [
-        f.name for f in output_folder.glob('*.csv') if incomplete in str(f)
-    ]
+    candidates = list(output_folder.glob('*.csv'))
+    return propose_from_list(candidates, incomplete)
+
+
+def get_conda_envs(ctx, args, incomplete):
+    "Auto-completion for env-names."
+    _, all_envs = list_conda_envs()
+    return propose_from_list(all_envs, incomplete)