+++ benchopt/tests/test_cli.py
@@ -8,12 +8,15 @@ from benchopt.plotting import PLOT_KINDS
 from benchopt.utils.stream_redirection import SuppressStd
 
 
+from benchopt.tests import CaptureRunOutput
 from benchopt.tests import SELECT_ONE_PGD
 from benchopt.tests import SELECT_ONE_SIMULATED
+from benchopt.tests import DUMMY_BENCHMARK
 from benchopt.tests import DUMMY_BENCHMARK_PATH
 
 
 from benchopt.cli.main import run
+from benchopt.cli.main import install
 from benchopt.cli.process_results import plot
 from benchopt.cli.helpers import check_install
 
@@ -38,46 +41,6 @@ class TestCheckInstallCmd:
         pgd_solver = DUMMY_BENCHMARK_PATH / 'datasets' / 'invalid.py'
         with pytest.raises(FileNotFoundError, match=r'invalid.py'):
             check_install([str(pgd_solver.resolve()), 'Dataset'], 'benchopt')
-
-
-class CaptureRunOutput(object):
-    """Context to capture run cmd output and files.
-    """
-
-    def __init__(self):
-        self.out = SuppressStd()
-        self.output = None
-        self.result_files = []
-
-    def __enter__(self):
-        self.output = None
-        self.result_files = []
-
-        # Redirect the stdout/stderr fd to temp file
-        self.out.__enter__()
-        return self
-
-    def __exit__(self, type, value, traceback):
-        self.out.__exit__(type, value, traceback)
-        self.output = self.out.output
-
-        # Make sure to delete all the result that created by the run command.
-        self.result_files = re.findall(
-            r'Saving result in: (.*\.csv)', self.output
-        )
-        if len(self.result_files) >= 1:
-            for result_file in self.result_files:
-                Path(result_file).unlink()
-
-        if type is not None:
-            print(self.output)
-
-    def check_output(self, pattern, repetition=None):
-        matches = re.findall(pattern, self.output)
-        if repetition is None:
-            assert len(matches) > 0, self.output
-        else:
-            assert len(matches) == repetition, self.output
 
 
 class TestRunCmd:
@@ -108,8 +71,8 @@ class TestRunCmd:
 
         out.check_output('Simulated', repetition=1)
         out.check_output('Dummy Sparse Regression', repetition=1)
-        out.check_output(r'Python-PGD\[step_size=1\]', repetition=3)
-        out.check_output(r'Python-PGD\[step_size=1.5\]', repetition=0)
+        out.check_output(r'Python-PGD\[step_size=1\]:', repetition=3)
+        out.check_output(r'Python-PGD\[step_size=1.5\]:', repetition=0)
 
         # Make sure the results were saved in a result file
         assert len(out.result_files) == 1, out.output
@@ -125,13 +88,14 @@ class TestRunCmd:
         out.check_output(f'conda activate {test_env_name}')
         out.check_output('Simulated', repetition=1)
         out.check_output('Dummy Sparse Regression', repetition=1)
-        out.check_output(r'Python-PGD\[step_size=1\]', repetition=3)
-        out.check_output(r'Python-PGD\[step_size=1.5\]', repetition=0)
+        out.check_output(r'Python-PGD\[step_size=1\]:', repetition=3)
+        out.check_output(r'Python-PGD\[step_size=1.5\]:', repetition=0)
 
         # Make sure the results were saved in a result file
         assert len(out.result_files) == 1, out.output
 
     def test_benchopt_caching(self):
+        # Check that the computation caching is working properly.
 
         n_rep = 2
         run_cmd = [str(DUMMY_BENCHMARK_PATH), '-l', '-d', SELECT_ONE_SIMULATED,
@@ -147,7 +111,7 @@ class TestRunCmd:
         with CaptureRunOutput() as out:
             run(run_cmd, 'benchopt', standalone_mode=False)
 
-        out.check_output(r'Python-PGD\[step_size=1\]',
+        out.check_output(r'Python-PGD\[step_size=1\]:',
                          repetition=1)
 
         # Make sure that -f option forces the re-run for the solver
@@ -155,8 +119,61 @@ class TestRunCmd:
         with CaptureRunOutput() as out:
             run(run_cmd, 'benchopt', standalone_mode=False)
 
-        out.check_output(r'Python-PGD\[step_size=1\]',
+        out.check_output(r'Python-PGD\[step_size=1\]:',
                          repetition=2*n_rep+1)
+
+
+class TestInstallCmd:
+
+    @pytest.mark.parametrize('invalid_benchmark, match', [
+        ('invalid_benchmark', "Path 'invalid_benchmark' does not exist."),
+        ('.', "The folder '.' does not contain `objective.py`")],
+        ids=['invalid_path', 'no_objective'])
+    def test_invalid_benchmark(self, invalid_benchmark, match):
+        with pytest.raises(click.BadParameter, match=match):
+            install([invalid_benchmark], 'benchopt', standalone_mode=False)
+
+    def test_invalid_dataset(self):
+        with pytest.raises(click.BadParameter, match=r"invalid_dataset"):
+            install([str(DUMMY_BENCHMARK_PATH), '-d', 'invalid_dataset',
+                     '-s', 'pgd'], 'benchopt', standalone_mode=False)
+
+    def test_invalid_solver(self):
+        with pytest.raises(click.BadParameter, match=r"invalid_solver"):
+            install([str(DUMMY_BENCHMARK_PATH), '-s', 'invalid_solver'],
+                    'benchopt', standalone_mode=False)
+
+    def test_existing_empty_env(self, empty_env_name):
+        msg = (
+            f"`benchopt` is not installed in existing env '{empty_env_name}'"
+        )
+        with pytest.raises(RuntimeError, match=msg):
+            install([str(DUMMY_BENCHMARK_PATH), '--env-name', empty_env_name],
+                    'benchopt', standalone_mode=False)
+
+    def test_benchopt_install(self):
+        with CaptureRunOutput() as out:
+            install(
+                [str(DUMMY_BENCHMARK_PATH), '-d', SELECT_ONE_SIMULATED, '-s',
+                 SELECT_ONE_PGD, '-y'], 'benchopt', standalone_mode=False
+            )
+
+        out.check_output(f"Installing '{DUMMY_BENCHMARK.name}' requirements")
+        out.check_output("already available\n", repetition=2)
+
+    def test_benchopt_install_in_env(self, test_env_name):
+        with CaptureRunOutput() as out:
+            install(
+                [str(DUMMY_BENCHMARK_PATH), '-d', SELECT_ONE_SIMULATED, '-s',
+                 SELECT_ONE_PGD, '--env-name', test_env_name],
+                'benchopt', standalone_mode=False
+            )
+
+        out.check_output(
+            f"Installing '{DUMMY_BENCHMARK.name}' requirements")
+        out.check_output(
+            f"already available in '{test_env_name}'\n", repetition=2
+        )
 
 
 class TestPlotCmd: