+++ benchopt/utils/misc.py
@@ -1,3 +1,5 @@
+
+
 def get_benchopt_requirement_line():
     """Specification for pip requirement to install benchopt in conda env.